requirements.txt
beautifulsoup4
fake-useragent
js2py
python-telegram-bot
requests
